refactor: RecruitingManagePage, RecruitingDetailPage 컴포넌트 정리
