+++ src/components/RecruitingDetail/detail/ApplicantCard.jsx
