+++ src/components/RecruitingDetail/detail/ApplicantListSection.css
