+++ src/components/RecruitingDetail/detail/RecruitingInfoSection.css
