+++ src/components/RecruitingDetail/detail/RecruitingInfoSection.jsx
