+++ src/components/RecruitingDetail/detail/index.js
