+++ src/components/RecruitingDetail/index.js
