+++ src/components/RecruitingDetail/stats/index.js
