--- src/components/pages/RecruitingManagement/components/index.js
@@ -1,4 +1,0 @@
-// RecruitingManagement Page Components
-export { default as SearchFilter } from './SearchFilter';
-export { default as StatsGrid } from './StatsGrid';
-export { default as RecruitingList } from './RecruitingList';
