+++ src/pages/RecruitingDetailPage.jsx
@@ -3,10 +3,10 @@ import { useParams, useNavigate } from 'react-router-dom';
 import { RecruitingDetailProvider, useRecruitingDetail } from '../contexts/RecruitingDetailContext.jsx';
 import { EvaluationProvider, useEvaluation } from '../contexts/EvaluationContext.jsx';
 import { AdminHeader } from '../components/admin';
-import { StatsSection } from '../components/recruiting/stats';
-import { ApplicantModal, EmailModal, BulkStatusChangeModal } from '../components/recruiting/modals';
+import { StatsSection } from '../components/RecruitingDetail/stats';
+import { ApplicantModal, EmailModal, BulkStatusChangeModal } from '../components/RecruitingDetail/modals';
 import DeleteConfirmModal from '../components/common/DeleteConfirmModal';
-import { RecruitingInfoSection, RecruitingHeader, ApplicantListSection } from '../components/recruiting/detail';
+import { RecruitingInfoSection, RecruitingHeader, ApplicantListSection } from '../components/RecruitingDetail/detail';
 import { RECRUITMENT_CONFIG, SORT_OPTIONS } from '../constants/recruitment';
 import { ROUTES } from '../constants/routes';
 import { sortApplicants } from '../utils/sort';
