+++ src/pages/RecruitingManagePage.jsx
@@ -5,7 +5,9 @@ import AdminCodeModal from '../features/admin/AdminCodeModal';
 import AdminCodeResultModal from '../features/admin/AdminCodeResultModal';
 import AdminManageModal from '../features/admin/AdminManageModal';
 import CreateRecruitingModal from '../features/admin/CreateRecruitingModal';
-import { SearchFilter, StatsGrid, RecruitingList } from '../components/pages/RecruitingManagement/components';
+import SearchFilter from '../components/pages/RecruitingManagement/SearchFilter';
+import StatsGrid from '../components/pages/RecruitingManagement/StatsGrid';
+import RecruitingList from '../components/pages/RecruitingManagement/RecruitingList';
 import Pagination from '../components/common/Pagination';
 import { useRecruitingManagement } from '../hooks/legacy/useRecruitingManagement';
 import { useModalManagement } from '../hooks/legacy/useModalManagement';
